+++ subsys/net/lib/aws_fota/Kconfig
@@ -13,6 +13,19 @@ menuconfig AWS_FOTA
 
 if AWS_FOTA
 
+config AWS_FOTA_MODEM_FW_UPDATE_AUTO_VALIDATE
+	bool "Auto validate modem firwmare update"
+	depends on NRF_MODEM_LIB_SYS_INIT
+	default y
+	help
+	  Auto validate the initialization of the modem firmware before
+	  the application (main thread) gets to run.
+	  During modem FOTA the modem needs to be reinitialized / rebooted twice.
+	  Once to apply the update and once to run the new image. During this process
+	  any socket / AT commands to the modem will fail. Enabling this option ensures that the
+	  the modem is in a healthy state before the application is allowed to run.
+	  This option will cause a reboot if required by the modem after DFU.
+
 config AWS_FOTA_PAYLOAD_SIZE
 	int "MQTT payload reception buffer size for AWS IoT Jobs messages"
 	default 1350
